+++ GPTifier/src/help_messages.cpp
@@ -1,4 +1,4 @@
-#include "commands.hpp"
+#include "help_messages.hpp"
 
 #include <iostream>
 #include <string>
