+++ GPTifier/src/main.cpp
@@ -1,5 +1,5 @@
 #include "command_run.hpp"
-#include "commands.hpp"
+#include "help_messages.hpp"
 #include "params.hpp"
 
 #include <iostream>
